# docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: Runic API
  description: |
    Firewall rule management system API.
    
    ## Authentication
    Most endpoints require Bearer token authentication. Include the token in the Authorization header:
    ```
    Authorization: Bearer <token>
    ```
    
    ## Error Responses
    All endpoints may return error responses in the following format:
    ```json
    {
      "error": "error message"
    }
    ```
    
    Common HTTP status codes:
    - 400: Bad Request - Invalid input
    - 401: Unauthorized - Missing or invalid token
    - 403: Forbidden - Insufficient permissions
    - 404: Not Found - Resource not found
    - 409: Conflict - Resource in use
    - 500: Internal Server Error
  version: 1.0.0
  contact:
    name: Runic Support
  license:
    name: MIT
servers:
  - url: http://localhost:8080/api/v1
    description: Local development server
  - url: /api/v1
    description: Production server (relative path)
...
paths:
...
  /auth/refresh:
    post:
      tags:
      - Auth
      summary: Refresh access token
      description: |
        Refreshes the access token using a valid refresh token.
        The refresh token is read from cookies.
      operationId: refreshToken
      responses:
        '200':
          description: Token refreshed successfully
          headers:
            Set-Cookie:
              schema:
                type: string
                example: "runic_access_token=...; HttpOnly; Path=/"
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "token refreshed"
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /auth/me:
    get:
      tags:
      - Auth
      summary: Get current user info
      description: Returns the authenticated user's profile information.
      operationId: getMe
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserInfo'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalServerError'
      security:
      - BearerAuth: []
...
  /users:
    get:
      tags:
      - Users
      summary: List all users
      description: Returns a list of all users. Requires admin role.
      operationId: listUsers
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/User'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '500':
          $ref: '#/components/responses/InternalServerError'
      security:
      - BearerAuth: []
    post:
      tags:
      - Users
      summary: Create a new user
      description: Creates a new user. Requires admin role.
      operationId: createUser
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateUserRequest'
      responses:
        '201':
          description: User created
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: integer
                    format: int64
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '500':
          $ref: '#/components/responses/InternalServerError'
      security:
      - BearerAuth: []
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      description: Bearer token authentication
  schemas:
...
    UserInfo:
      type: object
      properties:
        username:
          type: string
        role:
          type: string
          enum: [admin, editor, viewer]

    # System Info
...
    User:
      type: object
      properties:
        id:
          type: integer
          format: int64
        username:
          type: string
        email:
          type: string
          format: email
        role:
          type: string
          enum: [admin, editor, viewer]
        created_at:
          type: string
          format: date-time
    CreateUserRequest:
      type: object
      required:
      - username
      - password
      properties:
        username:
          type: string
          maxLength: 255
        password:
          type: string
          format: password
          minLength: 8
        email:
          type: string
          format: email
        role:
          type: string
          enum: [admin, editor, viewer]
...
    Error:
      type: object
      properties:
        error:
          type: string
  responses:
    BadRequest:
      description: Bad Request - Invalid input
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Unauthorized:
      description: Unauthorized - Missing or invalid token
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Forbidden:
      description: Forbidden - Insufficient permissions
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
    InternalServerError:
      description: Internal Server Error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'